imperative mode

+++ compile_libimobiledevice_imperatively.sh
@@ -1,30 +1,48 @@
 # This file is part of instructions in ibackupClient.sh and is intended to be run on a Raspberry Pi.
-# From https://gist.github.com/kfatehi/8922430
+# Based on https://gist.github.com/kfatehi/8922430
 
-sudo apt-get update
-sudo apt-get -y upgrade
-sudo apt-get -y install vim tmux git build-essential libxml2-dev python2.7 python2.7-dev fuse libtool autoconf libusb-1.0-0-dev libfuse-dev
+set -e
 
-mkdir iphone_libs && cd iphone_libs
+if [ ! -e iphone_libs ]; then
+    sudo apt-get update
+    sudo apt-get -y upgrade
+    sudo apt-get -y install vim tmux git build-essential libxml2-dev python2.7 python2.7-dev fuse libtool autoconf libusb-1.0-0-dev libfuse-dev
 
-git clone https://github.com/libimobiledevice/libplist.git
-git clone https://github.com/libimobiledevice/libusbmuxd.git
-git clone https://github.com/libimobiledevice/usbmuxd.git
-git clone https://github.com/libimobiledevice/libimobiledevice.git
-git clone https://github.com/libimobiledevice/ifuse.git
+    mkdir iphone_libs && cd iphone_libs
 
-export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig
-echo "export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig" | sudo tee -a /etc/bash.bashrc
+    git clone https://github.com/libimobiledevice/libplist.git
+    git clone https://github.com/libimobiledevice/libusbmuxd.git
+    git clone https://github.com/libimobiledevice/usbmuxd.git
+    git clone https://github.com/libimobiledevice/libimobiledevice.git
+    git clone https://github.com/libimobiledevice/ifuse.git
+    git clone https://github.com/libimobiledevice/libimobiledevice-glue.git
+fi
 
-cd libplist && ./autogen.sh && make && sudo make install && cd ..
+if [ -z "$(grep -F "export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig" /etc/bash.bashrc || true)" ]; then # Add if needed:
+    export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig
+    echo "export PKG_CONFIG_PATH=/usr/local/lib/pkgconfig" | sudo tee -a /etc/bash.bashrc
+fi
+
+if [ ! -e /usr/local/lib/libplist-2.0.a ]; then
+    cd libplist && ./autogen.sh && make && sudo make install && cd ..
+fi
+cd libimobiledevice-glue && ./autogen.sh && make && sudo make install && cd ..
+exit
 cd libusbmuxd && ./autogen.sh && make && sudo make install && cd ..
 cd libimobiledevice && ./autogen.sh && make && sudo make install && cd ..
 cd usbmuxd && ./autogen.sh && make && sudo make install && cd ..
 cd ifuse && ./autogen.sh && make && sudo make install && cd ..
 
-groupadd -g 140 usbmux &>/dev/null
-useradd -c 'usbmux user' -u 140 -g usbmux -d / -s /sbin/nologin usbmux &>/dev/null
-passwd -l usbmux &>/dev/null
+if [ -z "$(getent group usbmux || true)" ]; then # Add if needed:
+    # Add usbmux group
+    groupadd -g 140 usbmux
+    # Add usbmux user
+    useradd -c 'usbmux user' -u 140 -g usbmux -d / -s /sbin/nologin usbmux
+    # Make usbmux user have a locked password (can't be changed by the user)
+    passwd -l usbmux
+fi
 
-echo /usr/local/lib | sudo tee /etc/ld.so.conf.d/libimobiledevice-libs.conf
-sudo ldconfig
+if [ -z "$(grep -F "/usr/local/lib" /etc/ld.so.conf.d/libimobiledevice-libs.conf || true)" ]; then # Add if needed:
+    echo /usr/local/lib | sudo tee /etc/ld.so.conf.d/libimobiledevice-libs.conf
+    sudo ldconfig
+fi
